fix /user endpoint missing session migration
