+++ api/routes/user.js
@@ -6,7 +6,7 @@ async function getUser(req, res) {
   const {
     query: { forSignup },
     params: { username },
-    loggedInUser,
+    session: {user: loggedInUser},
   } = req;
   const isCheckingAvailability = typeof forSignup === 'string';
   if (!username) {
